+++ src/my_cpp_pkg/src/add_two_ints_server.cpp
@@ -20,7 +20,7 @@ private:
                             const example_interfaces::srv::AddTwoInts::Response::SharedPtr response)
     {
         response->sum = request->a + request->b;
-        RCLCPP_INFO(this->get_logger(), "%d + %d = %d", request->a, request->b, response->sum);
+        RCLCPP_INFO(this->get_logger(), "%d + %d = %d", (int)request->a, (int)request->b, (int)response->sum);
     }
 
     rclcpp::Service<example_interfaces::srv::AddTwoInts>::SharedPtr server_;
